# restaurante-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurante API",
        "description": "Documentación del backend (Laravel + Swagger).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios (paginado)",
                "operationId": "e9d4d8b419dbef04b5748fbf9ccf3881",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Usuario"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "4e9dbcbdaaaf9faab196a82d5cbaea23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario",
                                    "password",
                                    "nombre",
                                    "correo",
                                    "rol"
                                ],
                                "properties": {
                                    "usuario": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secreto123"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Demo",
                                        "nullable": true
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "rol": {
                                        "description": "admin|mesero|cocinero|cliente|empleado",
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Ver usuario por ID",
                "operationId": "e9ac05ddc4ea46d8848b012d7d92e3ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario existente",
                "operationId": "27ac608b6dd8b8e7c7c45b10c5210b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario": {
                                        "type": "string",
                                        "example": "admin2"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nuevo123"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Demo",
                                        "nullable": true
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin2@example.com"
                                    },
                                    "rol": {
                                        "type": "string",
                                        "example": "gerente"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario",
                "operationId": "ea700764859a665afd38bb39c2842f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Usuario": {
                "title": "Usuario",
                "description": "Modelo que representa un usuario del sistema",
                "required": [
                    "usuario",
                    "password",
                    "nombre",
                    "correo",
                    "rol"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "usuario": {
                        "type": "string",
                        "example": "admin"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Administrador"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Pérez",
                        "nullable": true
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@example.com"
                    },
                    "rol": {
                        "type": "string",
                        "example": "admin"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "restaurant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-12T15:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-12T15:30:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}